🐛 Fix freebsd/arm64 build break in disk-usage helper 🐛
The freebsd/arm64 Daily Build (commit 1011dd5dcc16) failed with:
"invalid operation: stat.Bavail * uint64(stat.Bsize) (mismatched types
int64 and uint64)". `unix.Statfs_t.Blocks`/`Bavail` are `uint64` on
linux but `int64` on freebsd, so multiplying them directly against a
`uint64(stat.Bsize)` conversion only compiled on some platforms.

- src/backup/diskusage_unix.go: explicitly convert both stat.Blocks and
  stat.Bavail to uint64 before multiplying by stat.Bsize, in both
  diskTotalBytes and diskFreeBytes, so the code compiles regardless of
  the platform-defined signedness of those fields

+++ src/backup/diskusage_unix.go
@@ -11,8 +11,11 @@ func diskTotalBytes(path string) (uint64, error) {
 	if err := unix.Statfs(path, &stat); err != nil {
 		return 0, err
 	}
-	// #nosec G115 -- Bsize/Blocks are platform-defined unsigned block counts
-	return stat.Blocks * uint64(stat.Bsize), nil
+	// #nosec G115 -- Bsize/Blocks are platform-defined block counts; Blocks
+	// is int64 on some platforms (e.g. freebsd/arm64) and uint64 on others,
+	// so both operands are explicitly converted rather than relying on
+	// untyped-constant promotion.
+	return uint64(stat.Blocks) * uint64(stat.Bsize), nil
 }
 
 // diskFreeBytes returns the space available to an unprivileged process on
@@ -23,6 +26,9 @@ func diskFreeBytes(path string) (uint64, error) {
 	if err := unix.Statfs(path, &stat); err != nil {
 		return 0, err
 	}
-	// #nosec G115 -- Bsize/Bavail are platform-defined unsigned block counts
-	return stat.Bavail * uint64(stat.Bsize), nil
+	// #nosec G115 -- Bsize/Bavail are platform-defined block counts; Bavail
+	// is int64 on some platforms (e.g. freebsd/arm64) and uint64 on others,
+	// so both operands are explicitly converted rather than relying on
+	// untyped-constant promotion.
+	return uint64(stat.Bavail) * uint64(stat.Bsize), nil
 }
